pyproject.toml
[project]
name = "hcc"
version = "0.1.0"
description = "A lightweight Python package for making HTTP requests with retry functionality"
authors = [
    { name = "Tamás Csertán", email = "tamas.csertan@holdonprojekt.hu" },
    { name = "Márk Sárvári", email = "mark@binarybush.dev" }
]
license = "GPL-3.0-or-later"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mypy>=1.15.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "ruff>=0.11.2",
    "types-pyyaml>=6.0.12.20250326",
    "types-requests>=2.32.0.20250328",
]

[tool.pytest.ini_options]
addopts = "tests --cov=hcc --cov-report=html -v"

